+++ changeLog.md
@@ -1,1 +1,2 @@
-1. 经过考虑和实际考察，决定移除`increase`比较方法
+1. `manifest.json`中`tag`支持填入`latest`
+2. 修改`semVer`比较方式代码